kell trying to update api

+++ src/main/java/com/demo/repository/AccountRepository.java
@@ -6,10 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, Integer> {
     @EntityGraph(attributePaths = "authorities")
     Optional<Account> findOneWithAuthoritiesByEmail(String email);
     @EntityGraph(attributePaths = "authorities")
     Optional<Account> findOneWithAuthoritiesByUsername(String username);
     Boolean existsByUsernameOrEmail(String username, String email);
+
 }
